0.0.8
新增上传列表
新增删除账号功能
新增上传等待列表
新增修改密码功能
新增找回密码功能
新增邮箱验证功能
新增修改邮箱功能
新增上传头像功能
新增个人设置页面
修改图片MD5为后端计算
修改重复图片直接返回图片信息
优化在游客状态下图片的信息
优化在登录状态下用户的数据

+++ server/src/utils/deleteImage.js
@@ -59,9 +59,11 @@ export const deleteImage = async (imageInfo, id) => {
       default:
         throw new Error(`未知的存储类型: ${type}`)
     }
-    // 删除本地缩略图
-    const thumbFilePath = path.join(process.cwd(), thumb)
-    await fs.unlink(thumbFilePath)
+    if (thumb) {
+      // 删除本地缩略图
+      const thumbFilePath = path.join(process.cwd(), thumb)
+      await fs.unlink(thumbFilePath)
+    }
     // 删除数据库记录
     await Image.deleteOne({ _id })
     // 删除相关的上传日志
